requirements.txt
streamlit
pandas
plotly
matplotlib
pyyaml